Moved all code from MediaPlayerActivity to VideoPlayerActivity
